+++ projects/yzy-ng/src/lib/date-selector/models/CalendarState.ts
@@ -1,0 +1,8 @@
+import { OptionModel } from './../../dropdown/models/OptionModel';
+import { DisplayDate } from './DisplayDate';
+
+export interface CalendarState {
+    id: number;
+    value: DisplayDate | OptionModel;
+    isOpen: boolean;
+}