+++ projects/yzy-ng/src/lib/field/enums/FieldTypes.ts
@@ -6,5 +6,7 @@ export enum FieldTypes {
     Checkbox,
     Number,
     Dropdown,
-    Radio
+    Radio,
+    Date,
+    File
 }