+++ projects/yzy-ng/src/lib/field/field.module.ts
@@ -1,3 +1,4 @@
+import { FileSelectorModule } from './../file-selector/file-selector.module';
 import { RadioModule } from './../radio/radio.module';
 import { CheckboxModule } from './../checkbox/checkbox.module';
 import { NgModule } from '@angular/core';
@@ -6,6 +7,7 @@ import { CommonModule } from '@angular/common';
 import { FieldComponent } from './field.component';
 import { DropdownModule } from '../dropdown/dropdown.module';
 import { FieldErrorComponent } from './field-error/field-error.component';
+import { DateSelectorModule } from '../date-selector/date-selector.module';
 
 @NgModule({
     declarations: [FieldComponent, FieldErrorComponent],
@@ -14,8 +16,10 @@ import { FieldErrorComponent } from './field-error/field-error.component';
         FormsModule,
         ReactiveFormsModule,
         DropdownModule,
+        DateSelectorModule,
         CheckboxModule,
-        RadioModule
+        RadioModule,
+        FileSelectorModule
     ],
     exports: [FieldComponent, FieldErrorComponent]
 })