+++ projects/yzy-ng/src/lib/field/models/FieldModel.ts
@@ -10,4 +10,6 @@ export interface FieldModel {
     options?: OptionModel[];
     isReadOnly?: boolean;
     validators?: ValidatorFn[];
+    isPlaceHolder?: boolean;
+    accept?: string;
 }