+++ projects/yzy-ng/src/lib/table/models/ColumnTypes.ts
@@ -2,5 +2,6 @@ export enum ColumnTypes {
     String,
     Number,
     Dropdown,
-    Boolean
+    Boolean,
+    Date
 }