+++ projects/yzy-ng/src/lib/table/table.module.ts
@@ -7,10 +7,18 @@ import { FormsModule } from '@angular/forms';
 import { CellComponent } from './cell/cell.component';
 import { PaginatorComponent } from './paginator/paginator.component';
 import { PipeModule } from '../pipes/pipe.module';
+import { DateSelectorModule } from '../date-selector/date-selector.module';
 
 @NgModule({
     declarations: [TableComponent, CellComponent, PaginatorComponent],
-    imports: [CommonModule, FormsModule, DropdownModule, PipeModule, CheckboxModule],
+    imports: [
+        CommonModule,
+        FormsModule,
+        DropdownModule,
+        DateSelectorModule,
+        PipeModule,
+        CheckboxModule
+    ],
     exports: [TableComponent]
 })
 export class TableModule {}