+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## [1.9.1](https://github.com/YzYpYzY/yzy-ng/compare/v1.9.0...v1.9.1) (2020-04-22)
+
+
+### Bug Fixes
+
+* **table:** remove margin checkbox ([5369559](https://github.com/YzYpYzY/yzy-ng/commit/53695597014828fc26d68958f778d8950a9f4cf6))
+
+# [1.9.0](https://github.com/YzYpYzY/yzy-ng/compare/v1.8.0...v1.9.0) (2020-04-22)
+
+
+### Features
+
+* add file field ([e1f9ef2](https://github.com/YzYpYzY/yzy-ng/commit/e1f9ef2b9e1e9b17161b000ff0035d4783e750ee))
+* **field:** manage placeholder and inline mode ([4c2255d](https://github.com/YzYpYzY/yzy-ng/commit/4c2255d47e57b16634a74f821886c44c0a70338c))
+
+# [1.8.0](https://github.com/YzYpYzY/yzy-ng/compare/v1.7.2...v1.8.0) (2020-04-20)
+
+
+### Features
+
+* add date selector ([257fe14](https://github.com/YzYpYzY/yzy-ng/commit/257fe145bbc2bea2469593ceaa1d0f354554e126))
+
+## [1.7.2](https://github.com/YzYpYzY/yzy-ng/compare/v1.7.1...v1.7.2) (2020-04-07)
+
+
+### Bug Fixes
+
+* change field default size strategy to take free space ([c25bd7d](https://github.com/YzYpYzY/yzy-ng/commit/c25bd7d87a086dcf8691fa416c0cfec452e13dd6))
+
 ## [1.7.1](https://github.com/YzYpYzY/yzy-ng/compare/v1.7.0...v1.7.1) (2020-04-07)